+++ src/main/java/adoptionshop/adoptionshop/Repository/AdoptionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdoptionRepository extends JpaRepository<AdoptionEntity, Integer> {
+public interface AdoptionRepository extends JpaRepository<AdoptionEntity, Long> {
 }
