+++ src/main/java/adoptionshop/adoptionshop/Repository/MedicalNotesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicalNotesRepository extends JpaRepository<MedicalNotesEntity, Integer> {
+public interface MedicalNotesRepository extends JpaRepository<MedicalNotesEntity, Long> {
 }
